Title: 최솟값 3개, Time: 58ms, Memory: 0MB, Status: Passed - CodeTree

+++ 250127/최솟값 3개/top-3-smallest-number.cpp
@@ -31,7 +31,7 @@ int main() {
 }
 
 void getAns(){
-    int tmp=1;
+    long long tmp=1;
     priority_queue<int> pq_tmp = pq;
     for (int i=0;i<3;i++){
         tmp *= pq_tmp.top();
